+++ Buildlease/Buildlease/ClientApp/src/components/cart/Cart.tsx
@@ -7,6 +7,7 @@ import MainContent from "../layout/MainContent";
 import CartFullView from "../views/CartFullView";
 import API from "../../API";
 import ProductOrderCard from "../cards/ProductOrderCard";
+import {Button, DatePicker} from "antd";
 
 export default function Cart() {
     
@@ -31,11 +32,28 @@ export default function Cart() {
                     <div  style={{
                         padding: '24px',
                     }}>
-                        <div>
-                            <h2>Actions:</h2>
+                        <div style={{
+                            margin: '0px',
+                            marginBottom: '16px',
+                        }}>
+                            <h2 style={{
+                                margin: '0px',
+                                marginBottom: '16px',
+                            }}>Actions:</h2>
+                            <Button block>Remove all</Button>
                         </div>
-                        <div>
-                            <h2>Options:</h2>
+                        <div style={{
+                            margin: '0px',
+                            marginBottom: '16px',
+                        }}>
+                            <h2 style={{
+                                margin: '0px',
+                                marginBottom: '16px',
+                            }}>Options:</h2>
+                            <div className='d-flex flex-row justify-content-between'>
+                                <DatePicker placeholder='Start date'/>
+                                <DatePicker placeholder='End date'/>
+                            </div>
                         </div>
                         <div>
                             <h2>Status:</h2>
