+++ Buildlease/Buildlease/ClientApp/src/components/order-history/OrderHistory.tsx
@@ -9,6 +9,7 @@ import MainContent from "../layout/MainContent";
 import OrderCard from '../cards/OrderCard';
 
 import API from "../../API";
+import {Button} from "antd";
 
 export default function OrderHistory() {
     
@@ -30,10 +31,29 @@ export default function OrderHistory() {
             </SubHeader>
             <div className='d-flex flex-row flex-grow-1'>
                 <SideMenu>
-                    
+                    <div style={{
+                        padding: '24px',
+                    }}>
+                        <div style={{
+                            marginBottom: '24px',
+                        }}>
+                            <h2 style={{
+                                margin: '0px',
+                                marginBottom: '16px',
+                            }}>Actions:</h2>
+                            <Button block>Show details for all</Button>
+                        </div>
+                        <div>
+                            <h2>Options:</h2>
+                        </div>
+                    </div>
                 </SideMenu>
                 <MainContent>
-                    {orderHistory?.map(orderView => <OrderCard OrderView={orderView} detailsShown={false}/>)}
+                    <div style={{
+                        padding: '24px',
+                    }}>
+                        {orderHistory?.map(orderView => <OrderCard OrderView={orderView} detailsShown={false}/>)}
+                    </div>
                 </MainContent>
             </div>
         </>
